+++ src/main/java/kr/modusplant/modules/auth/email/app/controller/EmailAuthController.java
@@ -7,9 +7,9 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import kr.modusplant.global.app.servlet.response.DataResponse;
-import kr.modusplant.modules.auth.email.model.request.EmailRequest;
-import kr.modusplant.modules.auth.email.model.request.VerifyEmailRequest;
-import kr.modusplant.modules.auth.email.service.AuthService;
+import kr.modusplant.modules.auth.email.app.http.request.EmailRequest;
+import kr.modusplant.modules.auth.email.app.http.request.VerifyEmailRequest;
+import kr.modusplant.modules.auth.email.app.service.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
